requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
python-dotenv==1.0.1
openai==1.35.15
httpx==0.27.0
